% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fillgap}
\alias{fillgap}
\title{fill gaps in times series with mean values}
\usage{
fillgap(d)
}
\arguments{
\item{d}{dataframe with times series}
}
\value{
dataframe
}
\description{
fill gaps in times series with mean values
}
