% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{importCSdata}
\alias{importCSdata}
\title{import data from raw data}
\usage{
importCSdata(filename, RetOpt = "data")
}
\arguments{
\item{filename}{name of file to be imported extension}

\item{RetOpt}{type of information to be imported. "info" import data heads and "data" import all data}
}
\value{
dataframe
}
\description{
import raw file data in Campbell .dat format
}
