% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tfif_days}
\alias{tfif_days}
\title{calculate thawing, freezing, isotermal and freezethaw days}
\usage{
tfif_days(timestamp, vt)
}
\arguments{
\item{vt}{vector with time series}

\item{timestam}{vector with date values}
}
\value{
list
}
\description{
calculate thawing, freezing, isotermal and freezethaw days
}
